Commit for Part 6: Static Files
This commit is cumulative of parts 2-6 for the files modified.
However, this message will only be relative to part 6.
The directory: '/polls/static' was created for the
purposes of storing files which need not be
manipulated by django, but rather only served as-is.
As per this tutorial, these include standard images and gifs.

A namespaced sub-directory called 'polls' makes django's
job easier when traversing directories following the
{%STATIC_URL%} tag's set path. The relative paths specified
in /polls/static/polls/style.css are appended to the value of
the STATIC_URL tag to locate multiple static files within multiple
static directories. The /polls/ url renders the static image
we specified in /polls/style.css within the section marked 'body'.