Part 2: Admin Site
In this section, I set up the developement server.
the 'manage.py createsuperuser' command prompted
for an email and password for the admin site which
is accessible on whatever IP I give as an argument
to 'manage.py runserver'.

admin.py was edited in this section to add
server-side changes from the admin page(in this case,
for the Question model). Additions were made
to allow increased functionality to the data that
can be manipulated and retrieved through the admin site.
This included date, time, and text fields.
Custom fields were added to include whether or not
if the Question added was published recently.
List display was added to show date-times
in a listed format on the side of the page.
It includes time-based search filters for listing Questions.

A 'DIRS' option was added to settings.py to indicate where
Django should look for templates for displaying its pages
in a certain format.