Part 3: Views
A view was added to views.py for the polls index.
This displays whenver a certain URL is visited on
the server (whichever is mapped to it). In polls/urls.py,
a url for the index view was added to the urlpatterns array.
Views for details, results, and votes were created to indicate
where the site should redirect should someone participate
in the polls app. Question leads to details which leads
to votes(options) which leads to the results page.
The url's for the corresponding views were added as well.
An HTTPRESPONSE output was added to the index view to
respond to the http request(from the user going to that url) by
calling the static html file that corresponds to that
particular view.

404 Error handling was added to the Question view in the event
someone tries accessing a question's url that doesn't exist.

These html files are stored in polls/templates/polls/nameofview.html.