Part 4: Templates, Urls, Views
In part 4, the code written previously was modified
to be simpler via more efficient methods in our views.py.
detail.html, index.html, and results.html are modified
to better display their corresponding views.

An HTML <form> element was added to detail.html. This
included various methods and loops to adjust
functionality within the individual pages of the webapp.
The voting buttons on the Question detail view use
called a POST method, which changes the code on the server
when a vote goes through. In this case, the database is altered
to increase the total amount of votes a question
has per associated choice.

I adjusted the urls.py to insure that we can access our
polls app via expected nomenclature: /polls/Question1/vote
This is an example of question 1's vote page url.
We adjusted the 'vote' function in views.py to check for
non-existent questions and to make sure that a correct
HttpResponse is delivered as a result of voting.

The results template was edited to display the results
in a basic format with a header and each choice with
its corresponding frequency.

Views.py was also edited to include generic views
instead of the basic one views I implemented previously.
These views are: ListView and DetailView