Part 5: Testing
In this section, polls/tests.py was edited to
address the bug with date-time fields for the
index view. Questions with dates of being created
in the future were listed as 'recently published',
To address this, the views.py was edited
to have was_published_recently() return
'now - datetime.timedelta(days=1) <= self.pub_date - now'

It now returns false if the publishing date is in the future.
This test was split up in tests.py next in order to have a
false-test and a true-test for old and future questions.
This is simply to make testing 'air-tight' and to make
sure everything is functioning as intended.

views.py was changed to  include the get_queryset() method
which will look for the 5 most recently published
questions and display them at the index of the app(addressed
the date-time bug here as well).

Methods to create questions and test them were added
to tests.py. If no questions exist, an appropriate
message will be displayed when someone tries accessing
the polls. A few more tests were added to test if
only past questions are displayed (verify bug fix).